+++ instagram-app/src/authentication/withAuthenticate.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const withAuthenticate = (Comp) => 
+    class extends React.Component {
+        render() {
+            return <Comp />
+        }
+    }
+
+
+export default withAuthenticate;
