+++ instagram-app/src/components/Login/Login.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './Login.css';
+
+const Login = () => {
+    return(
+        <div className='login-container'>
+            <form className='login-form'>
+                <input></input>
+            </form>
+        </div>
+    )
+}
